requirements.txt
pandas==2.2.3
numpy==1.26.4
requests==2.31.0
beautifulsoup4==4.12.2
great-expectations==0.15.50
lxml==5.3.2
sqlalchemy==2.0.23
python-dateutil==2.8.2